+++ test/message/test_ascii.py
@@ -1,6 +1,4 @@
 """Test transport."""
-import struct
-
 import pytest
 
 from pymodbus.message.ascii import MessageAscii
@@ -15,27 +13,6 @@ class TestMessageAscii:
         """Return message object."""
         return MessageAscii([1], False)
 
-
-    def test_check_LRC(self):
-        """Test check_LRC."""
-        data = struct.pack(">HHHH", 0x1234, 0x2345, 0x3456, 0x4567)
-        assert MessageAscii.check_LRC(data, 0x1C)
-
-    def test_check_noLRC(self):
-        """Test check_LRC."""
-        data = struct.pack(">HHHH", 0x1234, 0x2345, 0x3456, 0x4567)
-        assert not MessageAscii.check_LRC(data, 0x0C)
-
-    def test_compute_LRC(self):
-        """Test compute_LRC."""
-        data = struct.pack(">HHHH", 0x1234, 0x2345, 0x3456, 0x4567)
-        assert MessageAscii.compute_LRC(data) == 0x1c
-
-    def test_roundtrip_LRC(self):
-        """Test combined compute/check LRC."""
-        data = struct.pack(">HHHH", 0x1234, 0x2345, 0x3456, 0x4567)
-        assert MessageAscii.compute_LRC(data) == 0x1c
-        assert MessageAscii.check_LRC(data, 0x1C)
 
     @pytest.mark.parametrize(
         ("packet", "used_len", "res_id", "res"),
@@ -59,21 +36,6 @@ class TestMessageAscii:
         assert data == res
         assert not tid
         assert dev_id == res_id
-
-    @pytest.mark.parametrize(
-        ("data", "dev_id", "res_msg"),
-        [
-            (b'\x01\x05\x04\x00\x17', 1, b':010105040017DE\r\n'),
-            (b'\x03\x07\x06\x00\x73', 2, b':0203070600737B\r\n'),
-            (b'\x08\x00\x01', 3, b':03080001F4\r\n'),
-            (b'\x84\x01', 2, b':02840179\r\n'),
-        ],
-    )
-    def test_encode(self, frame, data, dev_id, res_msg):
-        """Test encode."""
-        msg = frame.encode(data, dev_id, 0)
-        assert res_msg == msg
-        assert dev_id == int(msg[1:3], 16)
 
     @pytest.mark.parametrize(
         ("data", "dev_id", "res_msg"),